+++ src/comp/metadata/encoder.rs
@@ -216,9 +216,9 @@ fn encode_discriminant(ecx: @encode_ctxt, ebml_w: ebml::writer, id: node_id) {
     ebml::end_tag(ebml_w);
 }
 
-fn encode_disr_val(_ecx: @encode_ctxt, ebml_w: ebml::writer, disr_val: int) {
+fn encode_disr_val(_ecx: @encode_ctxt, ebml_w: ebml::writer, disr_val: i32) {
     ebml::start_tag(ebml_w, tag_disr_val);
-    ebml_w.writer.write(str::bytes(int::to_str(disr_val,10u)));
+    ebml_w.writer.write(str::bytes(int::to_str(disr_val as int,10u)));
     ebml::end_tag(ebml_w);
 }
 
@@ -231,7 +231,7 @@ fn encode_tag_id(ebml_w: ebml::writer, id: def_id) {
 fn encode_tag_variant_info(ecx: @encode_ctxt, ebml_w: ebml::writer,
                            id: node_id, variants: [variant],
                            &index: [entry<int>], ty_params: [ty_param]) {
-    let disr_val = 0;
+    let disr_val = 0_i32;
     let i = 0;
     let vi = ty::tag_variants(ecx.ccx.tcx, {crate: local_crate, node: id});
     for variant: variant in variants {
@@ -253,7 +253,7 @@ fn encode_tag_variant_info(ecx: @encode_ctxt, ebml_w: ebml::writer,
         }
         encode_type_param_bounds(ebml_w, ecx, ty_params);
         ebml::end_tag(ebml_w);
-        disr_val += 1;
+        disr_val += 1_i32;
         i += 1;
     }
 }
